# perf: dynamic split-k for MLA