--- a/benchmarks/bench_deepseek_mla.py
+++ b/benchmarks/bench_deepseek_mla.py
@@ -59,21 +59,9 @@ def bench_deepseek_mla_decode(batch_size, seq_len, num_heads):
         q_nope.dtype,
         ckv.dtype,
     )
-    s = torch.cuda.Stream()
-    s.wait_stream(torch.cuda.current_stream())
-    with torch.cuda.stream(s):
-        for _ in range(3):
-            o = wrapper.run(q_nope, q_pe, ckv, kpe, return_lse=False)
-
-    torch.cuda.current_stream().wait_stream(s)
-
-    g = torch.cuda.CUDAGraph()
-    with torch.cuda.graph(g):
-        o = wrapper.run(q_nope, q_pe, ckv, kpe, return_lse=False)
 
     ms = triton.testing.do_bench(
-        # lambda: wrapper.run(q_nope, q_pe, ckv, kpe),
-        lambda: g.replay(),
+        lambda: wrapper.run(q_nope, q_pe, ckv, kpe),
         warmup=100,
         rep=1000,
     )
@@ -85,19 +73,6 @@ def bench_deepseek_mla_decode(batch_size, seq_len, num_heads):
 
 
 if __name__ == "__main__":
-    # bench_deepseek_mla_decode(768, 1024, 16)
-    # bench_deepseek_mla_decode(768, 1024, 32)
-    # bench_deepseek_mla_decode(768, 1024, 64)
-    # bench_deepseek_mla_decode(768, 2048, 16)
-    # bench_deepseek_mla_decode(768, 2048, 32)
-    # bench_deepseek_mla_decode(768, 2048, 64)
-    bench_deepseek_mla_decode(1, 1024, 16)
-    bench_deepseek_mla_decode(1, 1024, 32)
-    bench_deepseek_mla_decode(1, 1024, 64)
-    bench_deepseek_mla_decode(1, 2048, 16)
-    bench_deepseek_mla_decode(1, 2048, 32)
-    bench_deepseek_mla_decode(1, 2048, 64)
-    bench_deepseek_mla_decode(132, 1024, 16)
-    bench_deepseek_mla_decode(1, 32768, 64)
-    bench_deepseek_mla_decode(32, 32768, 16)
-    bench_deepseek_mla_decode(64, 32768, 16)
+    for seq_len in [1024, 2048, 4096, 8192, 16384, 32768]:
+        for batch_size in [1, 16, 32, 64]:
+            bench_deepseek_mla_decode(batch_size, seq_len, 16)