--- a/include/flashinfer/attention/mla_fa2.cuh
+++ b/include/flashinfer/attention/mla_fa2.cuh
@@ -619,39 +619,46 @@ __device__ __forceinline__ void finalize_m_(typename KTraits::AttentionVariant v
   }
 }
 
-template <uint32_t HEAD_DIM, typename IdType, typename DTypeO>
-__device__ void DevicePersistentMergeStates(IdType* merge_packed_offset_start,
-                                            IdType* merge_packed_offset_end, IdType* merge_indptr,
-                                            float* partial_o, float* partial_lse, DTypeO* final_o,
-                                            float* final_lse, const uint32_t o_stride_n,
-                                            const uint32_t o_stride_h,
+template <typename KTraits>
+__device__ void DevicePersistentMergeStates(typename KTraits::IdType* merge_packed_offset_start,
+                                            typename KTraits::IdType* merge_packed_offset_end,
+                                            typename KTraits::IdType* merge_indptr,
+                                            float* partial_o, float* partial_lse,
+                                            typename KTraits::DTypeO* final_o, float* final_lse,
+                                            const uint32_t o_stride_n, const uint32_t o_stride_h,
                                             const uint32_t cluster_tile_q,
                                             const uint_fastdiv& num_heads) {
-  constexpr uint32_t vec_size = 4;  // partial o has data type float
-  constexpr uint32_t num_threads_per_row = HEAD_DIM / vec_size;
-  const uint32_t cta_id = (blockIdx.z * gridDim.y + blockIdx.y) * gridDim.x + blockIdx.x;
+  constexpr uint32_t VEC_SIZE = 4;  // partial o has data type float
+  constexpr uint32_t NUM_THRS_PER_ROW = KTraits::HEAD_DIM_CKV / VEC_SIZE;
+  constexpr uint32_t ROWS_PER_ITERATION = (KTraits::NUM_THREADS) / NUM_THRS_PER_ROW;
+  const uint32_t cluster_id = blockIdx.y;
   const uint32_t thread_id = (threadIdx.z * blockDim.y + threadIdx.y) * blockDim.x + threadIdx.x;
+  const uint32_t offset_start = merge_packed_offset_start[cluster_id];
+  const uint32_t offset_end = merge_packed_offset_end[cluster_id];
+  const uint32_t partial_offset_start = merge_indptr[cluster_id];
+  const uint32_t partial_offset_end = merge_indptr[cluster_id + 1];
+  const uint32_t stride = offset_end - offset_start;
 #pragma unroll 1
-  for (uint32_t final_packed_offset =
-           merge_packed_offset_start[cta_id] + thread_id / num_threads_per_row;
-       final_packed_offset < merge_packed_offset_end[cta_id]; ++final_packed_offset) {
+  for (uint32_t local_packed_offset =
+           blockIdx.x * ROWS_PER_ITERATION + thread_id / NUM_THRS_PER_ROW;
+       local_packed_offset < stride; local_packed_offset += gridDim.x * ROWS_PER_ITERATION) {
+    uint32_t final_packed_offset = offset_start + local_packed_offset;
     uint32_t q, r;
     num_heads.divmod(final_packed_offset, q, r);
-    state_t<vec_size> st;
+    state_t<VEC_SIZE> st;
 #pragma unroll 2
-    for (uint32_t partial_packed_offset = merge_indptr[cta_id] + thread_id / num_threads_per_row;
-         partial_packed_offset < merge_indptr[cta_id + 1];
-         partial_packed_offset += cluster_tile_q) {
-      vec_t<float, vec_size> o_partial;
+    for (uint32_t partial_packed_offset = partial_offset_start + local_packed_offset;
+         partial_packed_offset < partial_offset_end; partial_packed_offset += stride) {
+      vec_t<float, VEC_SIZE> o_partial;
       float lse_partial;
-      o_partial.load(partial_o + partial_packed_offset * HEAD_DIM +
-                     (thread_id % num_threads_per_row) * vec_size);
+      o_partial.load(partial_o + partial_packed_offset * KTraits::HEAD_DIM_CKV +
+                     (thread_id % NUM_THRS_PER_ROW) * VEC_SIZE);
       lse_partial = partial_lse[partial_packed_offset];
       st.merge(o_partial, lse_partial, 1);
     }
     st.normalize();
-    st.o.cast_store(
-        final_o + (q * o_stride_n + r * o_stride_h + (thread_id % num_threads_per_row) * vec_size));
+    st.o.cast_store(final_o +
+                    (q * o_stride_n + r * o_stride_h + (thread_id % NUM_THRS_PER_ROW) * VEC_SIZE));
     if (final_lse) {
       final_lse[q * num_heads + r] = st.get_lse();
     }
@@ -679,7 +686,7 @@ __device__ __forceinline__ void write_o(typename KTraits::SharedStorage* smem_st
     for (uint32_t j = 0; j < 2; ++j) {
       uint32_t q_idx = (packed_offset + warp_idx_in_wg * 16 + 8 * j + lane_idx / 4) / num_heads;
       if (lane_idx % 4 == 0 && q_idx < q_len) {
-        partial_lse[warp_idx_in_wg * 16 + 8 * j + lane_idx / 4] =
+        partial_lse[(blockIdx.x * 4 + warp_idx_in_wg) * 16 + 8 * j + lane_idx / 4] =
             math::ptx_log2(d[j]) + float(m[j]);
       }
     }
@@ -691,11 +698,13 @@ __device__ __forceinline__ void write_o(typename KTraits::SharedStorage* smem_st
       for (uint32_t mma_d = 0; mma_d < NUM_MMA_D_CKV / 2; ++mma_d) {
         if (q_idx < q_len) {
           *reinterpret_cast<float2*>(
-              partial_o + (warp_idx_in_wg * 16 + 8 * j + lane_idx / 4) * HEAD_DIM_CKV +
+              partial_o +
+              ((blockIdx.x * 4 + warp_idx_in_wg) * 16 + 8 * j + lane_idx / 4) * HEAD_DIM_CKV +
               warpgroup_idx * (HEAD_DIM_CKV / 2) + mma_d * 16 + (lane_idx % 4) * 2) =
               *reinterpret_cast<float2*>(&o_frag[mma_d][j * 2]);
           *reinterpret_cast<float2*>(
-              partial_o + (warp_idx_in_wg * 16 + 8 * j + lane_idx / 4) * HEAD_DIM_CKV +
+              partial_o +
+              ((blockIdx.x * 4 + warp_idx_in_wg) * 16 + 8 * j + lane_idx / 4) * HEAD_DIM_CKV +
               warpgroup_idx * (HEAD_DIM_CKV / 2) + mma_d * 16 + 8 + (lane_idx % 4) * 2) =
               *reinterpret_cast<float2*>(&o_frag[mma_d][4 + j * 2]);
         }
@@ -814,7 +823,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
   const uint32_t kpe_stride_n = params.kpe_stride_n;
   const uint32_t o_stride_n = params.o_stride_n;
   const uint32_t o_stride_h = params.o_stride_h;
-  const uint32_t cluster_tile_q = blockDim.x * KTraits::CTA_TILE_Q;
+  const uint32_t cluster_tile_q = gridDim.x * KTraits::CTA_TILE_Q;
 
 #pragma unroll 1
   for (IdType work_idx = work_indptr[blockIdx.y]; work_idx < work_indptr[blockIdx.y + 1];
@@ -845,14 +854,15 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
             (CAUSAL ? min(kv_end, kv_len - q_len + (packed_qo_start + cluster_tile_q) / num_heads)
                     : kv_end),
             CTA_TILE_KV) -
-        1;
+        1 - (kv_start / CTA_TILE_KV);
 
-    int mask_tile_idx =
-        (CAUSAL ? min(kv_end, kv_len - q_len + packed_qo_start / num_heads) : kv_end) / CTA_TILE_KV;
+    // int mask_tile_idx =
+    //     (CAUSAL ? min(kv_end, kv_len - q_len + packed_qo_start / num_heads) : kv_end) /
+    //         CTA_TILE_KV -
+    //     (kv_start / CTA_TILE_KV);
+    int mask_tile_idx = 0;
 
-    int start_tile_idx = kv_start / CTA_TILE_KV;  // ceil_div(kv_start, CTA_TILE_KV);
     uint32_t block_iter_base = kv_indptr * block_size + kv_start;
-
     // last kv tile
     __syncthreads();
     uint32_t kv_bound = kv_indptr + (kv_len + block_size - 1) / block_size;  // ceil_div
@@ -863,7 +873,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
     cp_async::commit_group();
 #pragma unroll
     for (int stage_idx = 1; stage_idx < NUM_STAGES; ++stage_idx) {
-      if (kv_tile_idx - stage_idx >= start_tile_idx) {
+      if (kv_tile_idx - stage_idx >= 0) {
         load_kv<KTraits>(&smem_storage, ckv, kpe, kv_indices, ckv_stride_n, ckv_stride_page,
                          kpe_stride_n, kpe_stride_page, kv_bound,
                          block_iter_base + (kv_tile_idx - stage_idx) * CTA_TILE_KV, block_size,
@@ -874,7 +884,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
 
     // loop with mask
 #pragma unroll 1
-    for (; kv_tile_idx >= mask_tile_idx && kv_tile_idx > start_tile_idx; --kv_tile_idx) {
+    for (; kv_tile_idx >= mask_tile_idx && kv_tile_idx > 0; --kv_tile_idx) {
       cp_async::wait_group<NUM_STAGES - 1>();
       __syncthreads();
 
@@ -892,7 +902,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
       // compute sfm * v
       compute_mla_pv<KTraits>(&smem_storage, kv_tile_idx % NUM_STAGES, s_frag, d, o_frag);
 
-      if (kv_tile_idx - NUM_STAGES >= start_tile_idx) {
+      if (kv_tile_idx - NUM_STAGES >= 0) {
         __syncthreads();
         load_kv<KTraits>(&smem_storage, ckv, kpe, kv_indices, ckv_stride_n, ckv_stride_page,
                          kpe_stride_n, kpe_stride_page, kv_bound,
@@ -904,7 +914,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
 
     // loop without mask
 #pragma unroll 1
-    for (; kv_tile_idx + 1 > start_tile_idx + NUM_STAGES; --kv_tile_idx) {
+    for (; kv_tile_idx + 1 > NUM_STAGES; --kv_tile_idx) {
       cp_async::wait_group<NUM_STAGES - 1>();
       __syncthreads();
 
@@ -929,7 +939,7 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
 
     // last tiles
 #pragma unroll
-    for (; kv_tile_idx >= start_tile_idx; --kv_tile_idx) {
+    for (; kv_tile_idx >= 0; --kv_tile_idx) {
       // compute mla qk
       compute_mla_qk<KTraits>(&smem_storage, kv_tile_idx % NUM_STAGES, s_frag);
 
@@ -963,10 +973,10 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPagedAttentionKe
   grid.sync();
 
   // the second stage, merge partial outputs
-  DevicePersistentMergeStates<KTraits::HEAD_DIM_CKV>(
-      params.merge_packed_offset_start, params.merge_packed_offset_end, params.merge_indptr,
-      partial_o, partial_lse, final_o, final_lse, o_stride_n, o_stride_h, cluster_tile_q,
-      num_heads);
+  DevicePersistentMergeStates<KTraits>(params.merge_packed_offset_start,
+                                       params.merge_packed_offset_end, params.merge_indptr,
+                                       partial_o, partial_lse, final_o, final_lse, o_stride_n,
+                                       o_stride_h, cluster_tile_q, num_heads);
 }
 
 #define DISPATCH_SMEM_CONFIG(smem_limit_per_sm, NUM_STAGES, CTA_TILE_KV, QK_SHARD, ...) \