@yzh119 2025-02-17
Note that this PR uses f32 as the intermediate data type for output matrix in split-k.

Using 16bit data type (e.g. f16/bf16) might be faster, but will result in some accuracy loss, especially for bf16+long context.